requirements.txt
alembic==1.0.8
aniso8601==9.0.1
arrow==0.15.6
Babel==2.7.0
beautifulsoup4==4.8.2
bidict==0.21.2
blinker==1.4
Bootstrap-Flask==1.5.1
botocore==1.20.73
branca==0.4.1
cachelib==0.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
configparser==4.0.2
dnspython==1.16.0
docutils==0.15.2
dominate==2.4.0
eventlet==0.25.2
Flask==1.0.2
Flask-Babel==0.12.2
Flask-Bootstrap==3.3.7.1
Flask-HTTPAuth==4.4.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.4.0
Flask-Moment==0.9.0
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
folium==0.11.0
geographiclib==1.50
geojson==2.5.0
geopy==2.0.0
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.0.0
guess-language-spirit==0.5.3
gunicorn==20.0.4
idna==2.8
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
lxml==4.5.0
Mako==1.0.8
MarkupSafe==1.1.1
marshmallow==2.21.0
migrate==0.3.8
monotonic==1.5
numpy==1.18.5
orjson==3.4.8
packaging==20.8
pandas==1.0.4
phonenumbers==8.12.19
Pillow==7.0.0
Pint==0.16.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
PyJWT==1.7.1
pyowm==2.10.0
pyparsing==2.4.7
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
python-engineio==4.0.0
python-socketio==5.0.4
pytz==2019.3
requests==2.22.0
s3transfer==0.3.3
six==1.12.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
SQLAlchemy==1.3.2
stripe==2.48.0
twilio==6.53.0
typing-extensions==3.7.4.3
urllib3==1.25.7
visitor==0.1.3
Werkzeug==0.15.5
WTForms==2.2.1
WTForms-SQLAlchemy==0.1
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
